% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchingprocess-package.R
\docType{package}
\name{branchingprocess-package}
\alias{branchingprocess}
\alias{branchingprocess-package}
\title{branchingprocess: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
