% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qAny.R
\name{pFinalSize}
\alias{pFinalSize}
\title{Probability of outbreak extinction}
\usage{
pFinalSize(n, j, R, k)
}
\arguments{
\item{n}{number of initial cases in generation 0}

\item{j}{total outbreak size (>= n).}

\item{R}{mean of negative binomial offspring distribution}

\item{k}{dispersion of negative binomial offspring distribution}
}
\description{
qAny is the probability that n initial cases lead to an extinguished outbreak
of total size j after any number of transmission generations (j includes the
n initial cases)
}
\examples{
# With 5 initial individuals and negative binomial offspring distribution with mean R=0.2
# and dispersion k=0.1, gives the probability of outbreak extinction with a total number
# final outbreak size of exactly 5 to 20 inidividuals (including the initial 5).
pFinalSize(5, 5:20, R=0.2, k=0.1)

}
