% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFinalSizeAndGen.R
\name{pFinalSizeAndGen}
\alias{pFinalSizeAndGen}
\title{Joint probability of outbreak final size and number of transmission generations}
\usage{
pFinalSizeAndGen(g, n, j, R, k)
}
\arguments{
\item{g}{number of generations.}

\item{n}{number of initial cases}

\item{j}{final size}

\item{R}{reproduction number.}

\item{k}{dispersion parameter.}
}
\description{
Joint probability of outbreak final size and number of transmission generations
}
\author{
Damon Toth
}
