% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFinalSizeAndGenSwitch1.R
\name{pFinalSizeAndGenSwitch1}
\alias{pFinalSizeAndGenSwitch1}
\title{Joint probability of outbreak final size and number of transmission generations}
\usage{
pFinalSizeAndGenSwitch1(g, n, j, R0, k0, Rc, kc)
}
\arguments{
\item{g}{number of generations}

\item{n}{number of initial cases}

\item{j}{final size}

\item{R0}{mean of negative binomial offspring distribution from generation one}

\item{k0}{dispersion of negative binomial offspring distribution from generation one}

\item{Rc}{mean of negative binomial offspring distribution from generation two on}

\item{kc}{dispersion of negative binomial offspring distribution from generation two on}
}
\description{
Joint probability of outbreak final size and number of transmission generations
}
\author{
Damon Toth
}
