% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFinalSizeSwitch1.R
\name{pFinalSizeSwitch1}
\alias{pFinalSizeSwitch1}
\title{Probability of final outbreak size with offspring distribution parameters switched after generation one}
\usage{
pFinalSizeSwitch1(n, j, R0, k0, Rc, kc)
}
\arguments{
\item{n}{number of initial cases in generation 0}

\item{j}{total outbreak size (>= n).}

\item{R0}{mean of negative binomial offspring distribution from generation one}

\item{k0}{dispersion of negative binomial offspring distribution from generation one}

\item{Rc}{mean of negative binomial offspring distribution from generation two on}

\item{kc}{dispersion of negative binomial offspring distribution from generation two on}
}
\description{
pFinalSizeSwitch1 is the probability that n initial cases lead to an extinguished
outbreak of total size j after any number of transmission generations (j includes
the n initial cases)
}
\examples{
pFinalSizeSwitch1(5, 5:20, R0=2, k0=0.1, Rc=0.2, kc=0.1)

}
