% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFinalSizeSwitch2.R
\name{pFinalSizeSwitch2}
\alias{pFinalSizeSwitch2}
\title{Probability of final outbreak size with offspring distribution parameters switched after
two generations of transmission}
\usage{
pFinalSizeSwitch2(n, j, R0, k0, Rc, kc)
}
\arguments{
\item{n}{Number of initial cases in generation 0}

\item{j}{Total outbreak size (>= n).}

\item{R0}{Mean of negative binomial offspring distribution before the switch}

\item{k0}{Dispersion of negative binomial offspring distribution before the switch}

\item{Rc}{Mean of negative binomial offspring distribution after the switch}

\item{kc}{Dispersion of negative binomial offspring distribution before the switch}
}
\value{
The probability of the final outbreak size
}
\description{
pFinalSizeSwitch2 is the probability that n initial cases lead to an extinguished
outbreak of total size j after any number of transmission generations (j includes
the n initial cases), for a branching process with offspring distribution parameters
(R0,k0) during the first 2 transmission generations and (Rc,kc) during subsequent
generations
}
\examples{
#With 5 initial cases, the probability that the final outbreak size is 20
#(including the initial 5):
pFinalSizeSwitch2(n=5, j=20, R0=2, k0=0.1, Rc=0.2, kc=0.1)
}
