% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSizeAtGen.R
\name{pSizeAtGen}
\alias{pSizeAtGen}
\title{Probability that n initial cases lead to an outbreak that lasts at least g generations
of transmission AND has exactly j total cases after generation g}
\usage{
pSizeAtGen(g, n, j, R, k)
}
\arguments{
\item{g}{number of generations of transmission}

\item{n}{number of initial cases}

\item{j}{total size of outbreak after generation g}

\item{R}{reproduction number: mean of negative binomial offspring distribution}

\item{k}{dispersion parameter of negative binomial offspring distribution}
}
\description{
Probability that n initial cases lead to an outbreak that lasts at least g generations
of transmission AND has exactly j total cases after generation g
}
\author{
Damon Toth
}
