% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSizeAtGenSwitch1.R
\name{pSizeAtGenSwitch1}
\alias{pSizeAtGenSwitch1}
\title{Probability that n initial cases lead to an outbreak that lasts at least g generations
of transmission AND has exactly j total cases after generation g}
\usage{
pSizeAtGenSwitch1(g, n, j, R0, k0, Rc, kc)
}
\arguments{
\item{g}{number of generations of transmission}

\item{n}{number of initial cases}

\item{j}{total size of outbreak after generation g}

\item{R0}{basic reproduction number: mean of negative binomial offspring distribution from generation one}

\item{k0}{dispersion parameter of negative binomial offspring distribution from generation one}

\item{Rc}{control reproduction number: mean of negative binomial offspring distribution from generation two plus}

\item{kc}{dispersion parameter of negative binomial offspring distribution from generation two plus}
}
\description{
Probability that n initial cases lead to an outbreak that lasts at least g generations
of transmission AND has exactly j total cases after generation g
}
\author{
Damon Toth
}
