% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pExtinct.R
\name{pExtinct}
\alias{pExtinct}
\title{Probability that one initial case leads to an outbreak that eventually dies out (stochastic extinction)}
\usage{
pExtinct(R, k)
}
\arguments{
\item{R}{Reproduction number: mean of negative binomial offspring distribution}

\item{k}{Dispersion parameter of negative binomial offspring distribution}
}
\value{
The probability of outbreak extinction
}
\description{
Probability that one initial case leads to an outbreak that eventually dies out (stochastic extinction)
}
\examples{
# Probability that a single case leads to an extinct outbreak when the offspring distribution
# is negative binomial with mean 2 and dispersion parameter 0.5:
pExtinct(R=2,k=0.5)
}
\author{
Damon Toth
}
