% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFinalSizeAndGen.R
\name{pFinalSizeAndGen}
\alias{pFinalSizeAndGen}
\title{Joint probability of outbreak final size and number of transmission generations}
\usage{
pFinalSizeAndGen(g, n, j, R, k)
}
\arguments{
\item{g}{Number of generations.}

\item{n}{Number of initial cases}

\item{j}{Final size}

\item{R}{Reproduction number}

\item{k}{Dispersion parameter}
}
\value{
The joint probability of the final outbreak size and number of transmission generations
}
\description{
Joint probability of outbreak final size and number of transmission generations
}
\examples{
# Probability that 1 initial infection leads to an outbreak of final size 20 over exactly
# 3 generations of transmission:
pFinalSizeAndGen(g=3,n=1,j=20,R=0.8,k=0.1)
}
\author{
Damon Toth
}
