% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFinalSizeAndGen.R
\name{pFinalSizeAndGen}
\alias{pFinalSizeAndGen}
\title{Joint probability of outbreak final size and number of transmission generations}
\usage{
pFinalSizeAndGen(g, n, j, R, k)
}
\arguments{
\item{g}{Number of generations.}

\item{n}{Number of initial cases}

\item{j}{Final size}

\item{R}{Reproduction number}

\item{k}{Dispersion parameter}
}
\value{
The joint probability of the final outbreak size and number of transmission generations
}
\description{
Joint probability of outbreak final size and number of transmission generations
}
\author{
Damon Toth
}
