% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGen.R
\name{pGen}
\alias{pGen}
\title{Probability that one initial case leads to an outbreak lasting
less than g generations of transmission.}
\usage{
pGen(gMax, R, k)
}
\arguments{
\item{gMax}{Maximum number of generations.}

\item{R}{Reproduction number.}

\item{k}{Dispersion parameter.}
}
\value{
A vector of probabilities for each number of generations from 1 to gmax
}
\description{
Probability that one initial case leads to an outbreak lasting
less than g generations of transmission.
}
\author{
Damon Toth
}
