% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPgl.R
\name{pGen}
\alias{pGen}
\title{getPgl calculates \code{pgl[g]} for all g from 1 to gMax}
\usage{
pGen(gMax, R, k)
}
\arguments{
\item{gMax}{maximum number of generations.}

\item{R}{reproduction number.}

\item{k}{dispersion parameter.}
}
\description{
\code{pgl[g]} is the probability that one initial case leads to an outbreak lasting
less than g generations of transmission.
}
\author{
Damon Toth
}
