% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pNextGenSize.R
\name{pNextGenSize}
\alias{pNextGenSize}
\title{Probability of y total transmission directly from x independent infected individuals}
\usage{
pNextGenSize(x, y, R, k)
}
\arguments{
\item{x}{Number of infected individuals in generation n}

\item{y}{Number of total transmissions in generation n+1}

\item{R}{Mean of negative binomial offspring distribution}

\item{k}{Dispersion of negative binomial offspring distribution}
}
\value{
The probability of the given number of transmissions
}
\description{
Probability of y total transmission directly from x independent infected individuals
}
\examples{
# With 5 individuals in this generation, what is the probability of
# 0 to 15 transmissions in the next generation?
pNextGenSize(x=5, y=0:15, R=0.2, k=0.1)
}
